fast multiplication with u64 scalars
This multiplication implementation exits early (upon reaching the first
leading zero). Due to that it is useful for small scalars that fit into
u64.

Could be subject to side-channel attacks and is initially designed for
environments where the performance is critical while preserving constant
execution time or energy consumption is not.

Co-Authored-By: Giacomo Pasini <giacomo.pasini@iohk.io>